fix codecheck

+++ interfaces/kits/napi/file_access_module/napi_utils.cpp
@@ -28,12 +28,12 @@ namespace OHOS {
 namespace FileAccessFwk {
 int IsDirectory(const int64_t mode)
 {
-    if ((mode & DOCUMENT_FLAG_REPRESENTS_DIR) != DOCUMENT_FLAG_REPRESENTS_DIR) {
+    if ((static_cast<uint32_t>(mode) & DOCUMENT_FLAG_REPRESENTS_DIR) != DOCUMENT_FLAG_REPRESENTS_DIR) {
         HILOG_ERROR("current FileInfo is not dir");
         return ERR_INVALID_PARAM;
     }
 
-    if ((mode & DOCUMENT_FLAG_REPRESENTS_FILE) == DOCUMENT_FLAG_REPRESENTS_FILE) {
+    if ((static_cast<uint32_t>(mode) & DOCUMENT_FLAG_REPRESENTS_FILE) == DOCUMENT_FLAG_REPRESENTS_FILE) {
         HILOG_ERROR("file mode(%{public}" PRId64 ") is incorrect", mode);
         return ERR_INVALID_PARAM;
     }
